+++ src/components/Navbar.js
@@ -1,11 +1,13 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "./css/Navbar.module.css";
+import "../css/Navbar.module.css";
 
 function Navbar() {
   return (
     <nav>
       <Link to="/">Todo List</Link>
+      <Link to="new">New List</Link>
+
       <Link target="_blank" to="https://github.com/wessstt">
         Projects
       </Link>
