+++ src/components/TodoList.js
@@ -1,5 +1,6 @@
 import React from "react";
 import TodoListItem from "./TodoListItem";
+import PropTypes from "prop-types";
 
 const TodoList = ({ todoList, onRemoveTodo }) => {
   return (
@@ -9,5 +10,15 @@ const TodoList = ({ todoList, onRemoveTodo }) => {
       ))}
     </ul>
   );
+};
+
+TodoList.propTypes = {
+  todoList: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
+      title: PropTypes.string.isRequired,
+    })
+  ).isRequired,
+  onRemoveTodo: PropTypes.func.isRequired,
 };
 export default TodoList;
