+++ src/components/TodoListItem.js
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
-import styles from "./css/TodoListItem.module.css";
-import { ReactComponent as Delete } from "./svg/delete.svg";
+import styles from "../css/TodoListItem.module.css";
+import { ReactComponent as Delete } from "../svg/delete.svg";
+import PropTypes from "prop-types";
 
 const TodoListItem = ({ todo, onRemoveTodo }) => {
   const { title, id } = todo;
@@ -30,13 +31,21 @@ const TodoListItem = ({ todo, onRemoveTodo }) => {
         <button
           type="button"
           className={styles.buttonRemove}
-          onChange={() => onRemoveTodo(id)}
+          onClick={() => onRemoveTodo(id)}
         >
           <Delete height="25px" width="25px" />
         </button>
       </div>
     </li>
   );
+};
+
+TodoListItem.propTypes = {
+  todo: PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    title: PropTypes.string.isRequired,
+  }).isRequired,
+  onRemoveTodo: PropTypes.func.isRequired,
 };
 
 export default TodoListItem;
